refactor into a cache function
