+++ src/helpers/Generic.ts
@@ -20,3 +20,4 @@ export const handleSignal = (signal: string): void => {
   console.log(`Received ${signal}. Graceful shutdown...`);
   process.exit(0); // Exit the process
 };
+
